+++ helping_structure.hpp
@@ -1,0 +1,12 @@
+#ifndef HELPING_STRUCTURE_HPP_
+#define HELPING_STRUCTURE_HPP_
+
+struct coord
+{
+    unsigned int row:4;
+    unsigned int col:4;
+};
+
+
+
+#endif // !HELPING_STRUCTURE_HPP_
