+++ main.cpp
@@ -7,6 +7,9 @@
 #include <time.h>
 #include <stdlib.h>
 
+#include "puzzle_piece/puzzle_piece.hpp"
+#include "helping_structure.hpp"
+
 std::string decodedDate[51] = {"Jan","Feb","Mar","Apr","May","Jun", " ","\n",
                                "Jul","Aug","Sep","Oct","Nov","Dec", " ","\n",
                                 "01", "02", "03", "04", "05", "06","07","\n",
@@ -16,43 +19,6 @@ std::string decodedDate[51] = {"Jan","Feb","Mar","Apr","May","Jun", " ","\n",
                                 "29", "30", "31"};
 #define MAX_LOCATION 38
 #define DEBUG   0
-enum orientation_t {NOR_0=0 , NOR_90=1 , NOR_180=2 , NOR_270=3 , 
-                    MIR_0=4 , MIR_90=5 , MIR_180=6 , MIR_270=7};
-enum symmetry_t {NO_SYM = 0 , MIR_SYM=1 , ROT_SYM=2 , M_R_SYM = 3};
-struct coord
-{
-    unsigned int row:4;
-    unsigned int col:4;
-};
-
-class puzzle_piece
-{
-    private:
-    char m_pieceName;
-    coord m_filled[6];
-    coord m_dimensions;
-    orientation_t m_orientations[8];
-    unsigned int m_currentFilled:4;
-    unsigned int m_filledCount:4;
-    unsigned int m_currentOrientation:4;
-    unsigned int m_orientationsCount:4;
-    coord transformed_point();
-    public:
-
-    puzzle_piece();
-    puzzle_piece(char pieceName , std::string * strPtr , coord dimensions , symmetry_t shapeSymmetry);
-    void set_pieceName(char pieceName);
-    char get_pieceName();
-    void set_pieceShape(std::string * strPtr, coord dimensions);
-    void set_shapeSymmetry(symmetry_t shapeSymmetry);
-    bool get_next_orientation();
-    bool get_next_coord(coord &outCoord);
-    void set_currentOrientation(unsigned int index);
-    unsigned int get_currentOrientation();
-    coord get_dimensions();
-    void reset_shape();
-
-};
 
 std::string decode_date(std::string boardRep)
 {
@@ -115,7 +81,7 @@ struct backtrack
     puzzle_piece piece;
 };
 
-#define MAX_TIME 30*60*CLOCKS_PER_SEC
+#define MAX_TIME 1*60*CLOCKS_PER_SEC
 
 int main (int argc, char * argv[])
 {
@@ -319,183 +285,6 @@ int main (int argc, char * argv[])
     return 0;
 }
 
-void puzzle_piece::set_pieceName(char pieceName)
-{
-    this->m_pieceName = pieceName;
-}
-
-char puzzle_piece::get_pieceName()
-{
-    return this->m_pieceName;
-}
-
-void puzzle_piece::set_pieceShape(std::string * strPtr, coord dimensions)
-{
-    this->m_dimensions = dimensions;
-    this->m_filledCount =0;
-    if(strPtr != nullptr)
-    {
-        this->m_currentFilled = 0;
-        for (unsigned int i = 0; i < dimensions.row; i++)
-        {
-            for (unsigned int j = 0; j < dimensions.col; j++)
-            {
-                if(strPtr[i][j] !=' ')
-                {
-                    this->m_filled[this->m_currentFilled].row=i;
-                    this->m_filled[this->m_currentFilled].col=j;
-                    this->m_currentFilled++;
-                }
-            }
-        }
-        this->m_filledCount = this->m_currentFilled;
-    }
-    this->m_currentFilled = 0;
-}
-
-void puzzle_piece::set_shapeSymmetry(symmetry_t shapeSymmetry)
-{
-    this->m_currentOrientation = 2;
-    this->m_orientations[0] = NOR_0;
-    this->m_orientations[1] = NOR_90;
-    if(shapeSymmetry == NO_SYM || shapeSymmetry == MIR_SYM)
-    {
-        this->m_orientations[this->m_currentOrientation] = NOR_180;
-        this->m_orientations[this->m_currentOrientation+1] = NOR_270;
-        this->m_currentOrientation += 2;
-    }
-    if(shapeSymmetry == NO_SYM || shapeSymmetry == ROT_SYM)
-    {
-        for (unsigned int i = 0; i < this->m_currentOrientation; i++)
-        {
-            this->m_orientations[this->m_currentOrientation+i] = (orientation_t)(MIR_0 + this->m_orientations[i]);
-        }
-        this->m_currentOrientation *= 2;
-    }
-    this->m_orientationsCount = this->m_currentOrientation;
-    this->m_currentOrientation = 0;
-}
-
-bool puzzle_piece::get_next_orientation()
-{
-    bool result = false;
-    this->m_currentFilled = 0;
-    if(this->m_currentOrientation < this->m_orientationsCount-1)
-    {
-        this->m_currentOrientation++;
-        result = true;
-    }
-    return result;
-}
-
-void puzzle_piece::set_currentOrientation(unsigned int index)
-{
-    if( index < this->m_orientationsCount)
-    {
-        this->m_currentFilled = 0;
-        this->m_currentOrientation = index;
-    }
-}
-
-unsigned int puzzle_piece::get_currentOrientation()
-{
-    return this->m_currentOrientation;
-}
-
-coord puzzle_piece::transformed_point()
-{
-    coord result = {0,0};
-    switch(this->m_orientations[m_currentOrientation])
-    {
-        case NOR_0:
-        result.row = this->m_filled[this->m_currentFilled].row;
-        result.col = this->m_filled[this->m_currentFilled].col;
-        break;
-        
-        case NOR_90:
-        result.row = this->m_filled[this->m_currentFilled].col;
-        result.col = this->m_dimensions.row-this->m_filled[this->m_currentFilled].row-1;
-        break;
-
-        case NOR_180:
-        result.row = this->m_dimensions.row-this->m_filled[this->m_currentFilled].row-1;
-        result.col = this->m_dimensions.col-this->m_filled[this->m_currentFilled].col-1;
-        break;
-        
-        case NOR_270:
-        result.row = this->m_dimensions.col-this->m_filled[this->m_currentFilled].col-1;
-        result.col = this->m_filled[this->m_currentFilled].row;
-        break;
-
-        case MIR_0 :
-        result.row = this->m_filled[this->m_currentFilled].row;
-        result.col = this->m_dimensions.col-this->m_filled[this->m_currentFilled].col-1;
-        break;
-        
-        case MIR_90 :
-        result.row = this->m_dimensions.col-this->m_filled[this->m_currentFilled].col-1;
-        result.col = this->m_dimensions.row-this->m_filled[this->m_currentFilled].row-1;
-        break;
-
-        case MIR_180 :
-        result.row = this->m_dimensions.row-this->m_filled[this->m_currentFilled].row-1;
-        result.col = this->m_filled[this->m_currentFilled].col;
-        break;
-
-        case MIR_270 :
-        result.row = this->m_filled[this->m_currentFilled].col;
-        result.col = this->m_filled[this->m_currentFilled].row;
-        break;
-    }
-    return result;
-}
-
-bool puzzle_piece::get_next_coord(coord &outCoord)
-{
-     bool result = false;
-    if(this->m_currentFilled < this->m_filledCount)
-    {
-        outCoord = this->transformed_point();
-        this->m_currentFilled++;
-        result = true;
-    }
-    return result;
-}
-
-coord puzzle_piece::get_dimensions()
-{   
-    coord result = {0,0};
-    if(this->m_orientations[this->m_currentOrientation] & 1)//odd
-    {
-        result.row = this->m_dimensions.col;
-        result.col = this->m_dimensions.row;
-    }
-    else
-    {
-        result.row = this->m_dimensions.row;
-        result.col = this->m_dimensions.col;
-    }
-    return result;
-}
-
-void puzzle_piece::reset_shape()
-{
-    this->m_currentFilled = 0;
-    this->m_currentOrientation = 0;
-}
-
-puzzle_piece::puzzle_piece(char pieceName ,std::string * strPtr , coord dimensions , symmetry_t shapeSymmetry)
-{
-    this->set_pieceName(pieceName);
-    this->set_pieceShape(strPtr,dimensions);
-    this->set_shapeSymmetry(shapeSymmetry);
-    this->reset_shape();
-}
-
-puzzle_piece::puzzle_piece()
-{
-    puzzle_piece('a',nullptr,{0,0},NO_SYM);
-}
 
 puzzle_board::puzzle_board()
 {
